requirements/common.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
facepy==1.0.11
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.3.0
marshmallow-sqlalchemy==0.21.0
PyMySQL==0.7.11
python-dotenv==0.10.3
requests==2.23.0
six==1.13.0
SQLAlchemy==1.3.12
stripe==2.46.0
urllib3==1.25.8
Werkzeug==0.16.0
